requirements.txt
pydicom
numpy
pillow
matplotlib
PySide6
python-gdcm
py2app
